Added ESLint and Prettier for formatting and lint
Adding ESLint and Prettier as dev dependencies
with config files in the root of the project. Also
locally installed the ESLint and Prettier
VScode extensions. The ESLint extension actively
highlights and shows syntactical errors based on
the ESLint config file in the root of the
current project. The prettier extension allows
VScode to run the prettier formatter with a
keyboard command or through the VScode editor
command line. I already set prettier to be the
default formatter, so the normal command
ctrl+shift+i works.